parse

+++ input_arguments/printers.c
@@ -1,6 +1,6 @@
 #include "../cube.h"
 
-int ft_strlen(char *str)
+size_t  ft_strlen(char *str)
 {
     int i;
